pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling", "uv-dynamic-versioning" ]

[project]
name = "gitlabbot"
description = "Make comment on gitlab using `flux-local` diff outputs."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]
dependencies = [
    "flux-local>=7.1",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "pydantic-cli>=9.1",
    "pydantic-settings>=2.8.1",
    "python-gitlab>=5.6",
]

repository = "https://github.com/Alexsaphir/GitlabBot"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
typing = [
    "pyright>=1.1.396",
]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.pyproject-fmt]
indent = 4
max_supported_python = "3.13"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true
pattern = "default-unprefixed"

[tool.uv]
default-groups = [ 'dev', 'typing' ]
